simplify board with ByRole
