+++ src/main/scala/Role.scala
@@ -1,5 +1,8 @@
 package chess
 
+import cats.Functor
+import cats.Foldable
+
 sealed trait Role:
   val forsyth: Char
   lazy val forsythUpper: Char = forsyth.toUpper
@@ -57,3 +60,26 @@ object Role:
       case Rook   => Option(5)
       case Queen  => Option(9)
       case King   => None
+
+case class ByRole[A](pawn: A, knight: A, bishop: A, rook: A, queen: A, king: A):
+  def apply(role: Role): A = role match
+    case Pawn   => pawn
+    case Knight => knight
+    case Bishop => bishop
+    case Rook   => rook
+    case Queen  => queen
+    case King   => king
+
+  def values: List[A] = List(pawn, knight, bishop, rook, queen, king)
+
+object ByRole:
+  given Functor[ByRole] with
+    def map[A, B](byRole: ByRole[A])(f: A => B): ByRole[B] =
+      ByRole(
+        f(byRole.pawn),
+        f(byRole.knight),
+        f(byRole.bishop),
+        f(byRole.rook),
+        f(byRole.queen),
+        f(byRole.king)
+      )
